Don't treat the whole word as an URL
Changed re.search to re.match because .search scan through the whole string
looking for the regex. So, for example "asodijfosidwww.google.comaosdfoisad"
with .search will be treated as a whole URL.

With .match the complete word will be treated as an URL if the whole word
matches with the regular expression.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>