Show previous message sent on up and down arrow keys press
Pressing the up arrow key shows previously entered messages and the entry seems to lose focus
but pressing the up arrow key still works even if there's no visible indication of the entry gaining focus,
pressing the down arrow key at this point doesn't work as the entry doesn't have focus.

Pressing the down arrow key when the entry has focus works as expected and when they're no more messages to
show, the entry is empty for any new messages.

I'm not sure if this is a Gtk or Sugar bug but same thing doesn't happen in a standalone program so it is
most likely a Sugar problem.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>